randr: Properly delete old mode in multi-monitor configs
Using x11->width and x11->height to delete the old mode is wrong, since
this contains the total screen width and height, not the per monitor
width and height.

And we must not only delete the mode when replacing it with a new one,
but also when disabling the output.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>