randr: Make all sizes and coordinates signed
Xrandr itself has monitor sizes and coordinates as signed, and so does
most of our code, but not all of it. Making it all signed fixes a compiler
warning, and allows us to properly deal with the client sending negative
monitor coordinates.

Note: Unfortunately this does require some pointer casts when dealing with
the agent monitor-config message, as width and height are unsigned there :|

Signed-off-by: Hans de Goede <hdegoede@redhat.com>